+++ src/pages/pomodoro.tsx
@@ -1,11 +1,7 @@
-import WindowWrapper from "@/components/window";
-import { useTranslation } from "react-i18next";
+import PomodoroContent from "@/contents/pomodoro";
 
 export default function PomodoroPage(){
-     const {t} = useTranslation("pomodoro")
      return (
-          <WindowWrapper title={t("title")}>
-               Hello New Calm World, Goodbye Hellish world!
-          </WindowWrapper>
+          <PomodoroContent/>
      )
 }